- Fixed enums in simple fileds - Continued refactoring Array fields to Lists - Started fixing the AB-ETH and ADS driver after this change - Refactored the ADS driver to use byte[] instead of Lists of "int 8"

+++ plc4j/spi/src/main/java/org/apache/plc4x/java/spi/generation/ReadBufferXmlBased.java
@@ -111,7 +111,8 @@ public class ReadBufferXmlBased implements ReadBuffer, BufferCommons {
         hexString = hexString.substring(2);
         byte[] bytes = new byte[numberOfBytes];
         for (int i = 0; i < hexString.length(); i = i + 2) {
-            bytes[i / 2] = Byte.parseByte(hexString.substring(i, i + 2), 16);
+            // Without this hack, we can't parse values such as "0x80" into a byte.
+            bytes[i / 2] = (byte) Short.parseShort(hexString.substring(i, i + 2), 16);
         }
         return bytes;
     }
